fix: price rounding algorithm

+++ src/main/java/ru/mirea/megatracker/dto/CoinInfoDTO.java
@@ -4,6 +4,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 @Getter
 @Setter
@@ -11,10 +13,10 @@ public class CoinInfoDTO {
     private String name;
     private String ticker;
     private String iconUrl;
-    private double currentPrice;
-    private double deltaPrice;
-    private double deltaPricePercent;
+    private float currentPrice;
+    private float deltaPrice;
+    private float deltaPricePercent;
     private long marketCap;
-    private double highDay;
-    private double lowDay;
+    private float highDay;
+    private float lowDay;
 }
